Move fourhills into subpackage with imports to compensate
