+++ fourhills/__init__.py
@@ -1,0 +1,11 @@
+from .setting import Setting
+from .stats import StatBlock
+from .npc import Npc
+
+name = "fourhills"
+
+__all__ = [
+    "Npc",
+    "Setting",
+    "StatBlock",
+]
