+++ fourhills/text_utils.py
